pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "snippy-cli"
version = "1.0.0"
description = "A CLI tool for managing code snippets"
readme = "README.md"
authors = [{name = "Tejas Baid", email = "tejasbaid3@gmail.com"}]
requires-python = ">=3.6"
dependencies = ["pyperclip", "python-editor"]

[project.urls]
homepage = "https://github.com/TejasBaid/snippy-cli"

[project.scripts]
snip = "snippy.__main__:main"
